+++ lib/src/indicators/vol_ma.dart
@@ -147,7 +147,7 @@ class VolMaPaintObject<T extends VolMaIndicator>
       if (ret == null) continue;
 
       final text = formatNumber(
-        ret.val,
+        ret.val.toDecimal(),
         precision: state.curKlineData.req.precision,
         cutInvalidZero: true,
         prefix: '${param.label}: ',
